split modules, fix smaller bugs, add display name to configuration
