+++ opcpublisher/OpcSubscription.cs
@@ -1,0 +1,27 @@
+﻿
+using Opc.Ua.Client;
+using System.Collections.Generic;
+
+namespace OpcPublisher
+{
+    using static OpcApplicationConfiguration;
+
+    /// <summary>
+    /// Class to manage OPC subscriptions. We create a subscription for each different publishing interval
+    /// on an Endpoint.
+    /// </summary>
+    public class OpcSubscription
+    {
+        public List<OpcMonitoredItem> OpcMonitoredItems;
+        public int RequestedPublishingInterval;
+        public double PublishingInterval;
+        public Subscription OpcUaClientSubscription;
+
+        public OpcSubscription(int? publishingInterval)
+        {
+            RequestedPublishingInterval = publishingInterval ?? OpcPublishingInterval;
+            PublishingInterval = RequestedPublishingInterval;
+            OpcMonitoredItems = new List<OpcMonitoredItem>();
+        }
+    }
+}
